ApproxEq

Description

Provides a typeclass for for comparing shapes that are approxmately the same size and have the same position in 2-d space.

Documentation

class Show s => ApproxEq s

Minimal complete definition

approxEqual

approxEqual :: ApproxEq s => s -> s -> Float -> Bool

Are two shapes identical to within a given tolerance?

(=~) :: ApproxEq s => s -> s -> Bool

Are two shapes identical to within a given tolerance? Infix version where epsilon is fixed at 0.0001.

assertApproxEqual :: (ApproxEq s, HasCallStack) => s -> s -> Float -> Assertion

Helpful assertion for testing.

(@?~) :: ApproxEq s => s -> s -> Assertion

Helpful assertion for testing. Infix version where epsilon is fixed at 0.01.