Group

Description

A module for creating collections of homogenous shape types, optimizing shapes (combining paths when possible or removing duplicates), applying transformations to all of them, and rendering them as SVG elements.

Documentation

newtype Group shape

A type reprresenting a named group of some type of shape.

Constructors

Group [shape]

toLayer :: (Show a, SvgShape a, Transformable a) => String -> Group a -> Layer

Convert a group to a layer.

size :: Group a -> Int

Number of shapes in the group.

deduplicate :: (Ord t, ApproxEq t) => Group t -> Group t

Remove any duplicate shapes from this group

transformAndAppend :: (Show a, SvgShape a, Transformable a, Mergable a) => (a -> a) -> Group a -> Group a

optimizeGroup :: Mergable a => Group a -> Float -> Group a

Produce a new group where all shapes that can be merged have been, plus all shapes that couldn't be.

optimizeGroupAndLog :: Mergable a => Group a -> Float -> IO (Group a)

Produce a new group where all shapes that can be merged have been, plus all shapes that couldn't be, logging the result at the end.

toSvgN :: SvgShape s => Group s -> Svg

Wrap a group of shapes into an SVG g element, including a comment with the name of the group.

toSvgWithStyle :: SvgShape s => Group s -> StyleAttrs -> Svg

Wrap a group of shapes into an SVG g element, including a comment with the name of the group. Apply the given styles to the group.