Layer

Description

A module for creating collections of heterogenous shape types with uniform styling and rendering them as SVG elements.

Documentation

data ShapeLike

We use an existential type to wrap anything that meets our defintition of a shape. That is to say it can be shown, can be transformed according to methods in the Transformable typeclass, and can be converted to an SVG element. This includes arcs, circles, lines, and rectangles. Heterogenous shapes can all be placed in a single list and transformed via this type.

pattern MkShape :: a -> ShapeLike

data Layer

A layer is just a sequence of shape-like things with a name and style

shapes :: Layer -> [ShapeLike]

style :: Layer -> Maybe StyleAttrs

mkLayer :: (Show a, SvgShape a, Transformable a) => String -> [a] -> Layer

Layer constructor from a list of shapes: without style

mkLayerWithStyle :: (Show a, SvgShape a, Transformable a) => String -> [a] -> StyleAttrs -> Layer

Layer constructor from a list of shapes: with style

pack :: (Show a, SvgShape a, Transformable a) => a -> ShapeLike

A constructor to create a shape-like object from any shape.

(+:) :: (Show a, SvgShape a, Transformable a) => a -> Layer -> Layer infixr 5

Cons operation on a layer:

layer = mkLayer "layer-1" [shape1, shape3, shape3]

newlayer = shape+: layer

translate :: Layer -> Float -> Layer

Functions that mirror those in the Transformable typeclass, but act on a whole layer of shapes, rather than a single shape.

translateP :: Layer -> Point -> Layer

rotate :: Layer -> Point -> Float -> Layer

mirror :: Layer -> Point -> Point -> Layer

offset :: Layer -> Point -> Bool -> Layer