Line

Documentation

data Line

pattern Line :: Point -> Point -> Line

start :: Line -> Point

end :: Line -> Point

mkLine :: Float -> Float -> Float -> Float -> Line

Short-form constructor for a line.

asTuple :: Line -> (Float, Float, Float, Float)

Represent a line as a 4-tuple.

lineLength :: Line -> Float

Calculate the length of a line.

slope :: Line -> Either String Float

Calculate the slope of a line.

areParallel :: Line -> Line -> Float -> Bool

Are two lines parallel?

containsPoint :: Point -> Line -> Bool

Does the given point lie inside the given line segment?

overlappingSegments :: Line -> Line -> Bool

Are two line (segments) overlapping? (i.e. are collinear and the enpoint of one is contained within length of the other)