Point

Documentation

data Point

A point in 2-d space:

Point 3 4

pattern Point :: !Float -> !Float -> Point

xVal :: Point -> Float

Extract the x axis value from a point

yVal :: Point -> Float

Extract the y axis value from a point

fromFloat :: Float -> Point

Construct a point from a scalar value.

pointMap :: (Float -> Float) -> Point -> Point

Map a float -> float function over a point.

pointAsTuple :: Point -> (Float, Float)

Convert a point to a 2-tuple representation.

Operators for doing arithmetic on a point with a scalar float value

(|+|) :: Point -> Float -> Point

Add a scalar value to a point:

>>> Point 1 2 |+| 5
Point 6 7

(|*|) :: Point -> Float -> Point

Multiply a point by a scalar value:

>>> Point 1 2 |*| 5
Point 5 10

(|/|) :: Point -> Float -> Point

Divide a point by a scalar value:

>>> Point 10 10 |/| 5
Point 2 2

Geometric point manipulation functions

cross :: Point -> Point -> Float

dot :: Point -> Point -> Float

angleBetween :: Point -> Point -> Float

mag :: Point -> Float

magSquared :: Point -> Float

mirrorP :: Point -> Point -> Point -> Point

Mirror point a about a line through point p along vector v

rotateP :: Point -> Point -> Float -> Point

Rotate point a about a line through point p along vector t

cartesianProduct :: [Float] -> [Float] -> [Point]

Return the cartesian product of a set of coordinates on the X axis and a set of coordinates on the Y axis as a series of points.