Rectangle

Description

A type representing a parallelogram (doesn't just have to be a rectangle).

Documentation

data Rectangle

Constructors

Rectangle

Fields

topLeft :: Point

topRight :: Point

bottomRight :: Point

bottomLeft :: Point

mkRectangle :: Point -> Point -> Rectangle

Constructor for a rectangle; supply one point for the top left and one point for the bottom right.

mapPoints :: (Point -> Point) -> Rectangle -> Rectangle

Map a function over each point in the rectangle.

origin :: Rectangle -> Point

Midpoint of a rectangle.

asLines :: Rectangle -> [Line]

Convert a rectangle to a list of lines