Scene

Description

A Scene encapsualtes all SVG elements that will be rendered to a file.

Documentation

data Scene

pattern Scene :: Float -> Float -> StyleAttrs -> [Svg] -> Scene

width :: Scene -> Float

height :: Scene -> Float

style :: Scene -> StyleAttrs

elements :: Scene -> [Svg]

mkScene :: Float -> Float -> [Layer] -> Scene

Create a scene (without an associated style). Provide height, width (in inches) and a list of layers.

mkSceneWithStyle :: Float -> Float -> StyleAttrs -> [Layer] -> Scene

Create a scene (with an associated style). Provide height, width (in inches), style, and a list of layers.

svgDoc :: Scene -> Svg -> Svg

Given a scene and an SVG element representing it's contents, render the final scene.

emptyScene :: Float -> Float -> StyleAttrs -> Scene

addElement :: SvgShape s => Scene -> s -> Scene

renderScene :: IO Scene -> IO String

Render a scene to a string containin all SVG code.