Scenes.Registry

Description

A module for handling registration of scenes so they can be rendered by the CLI executable.

Documentation

getSceneOrExit :: String -> IO Scene

Return the named scene or exit the program with an error if it doesn't exist.

scenes :: Map String (IO Scene)

Register scenes and their CLI lable here.

scenesLabel :: String

Comma-separated list of all scene labels registered in thei module.