Shape

Description

Typeclasses for shapes

Documentation

class SvgShape s

Typeclass for shapes that can be converted to SVG elements

Minimal complete definition

toSvg

toSvg :: SvgShape s => s -> Svg

Convert a shape to a blaze-svg element.

class Transformable s

Typeclass for shapes that have transformation operations in cartesian space

Minimal complete definition

translate, rotate, mirror, offset

translate :: Transformable s => Point -> s -> s

Move a shape in space adding the given point to the X and Y axes.

translatePoints :: Transformable s => [Point] -> s -> [s]

Duplicate a shape in spcae across a list of points.

rotate :: Transformable s => Point -> Float -> s -> s

Rotate a shape about a line through point p by t radians

mirror :: Transformable s => Point -> Point -> s -> s

Mirror a shape about a line through point p along vector v

offset :: Transformable s => Point -> Bool -> s -> s

class Mergable s

Minimal complete definition

merge

merge :: Mergable s => s -> s -> Float -> Maybe s

If two shapes can be merged into one, returen the merged shape Makes the most sense for lines and arcs.

optimize :: Mergable s => [s] -> Float -> [s]

Merge all shapes in a given list that can be merged