Arc

Documentation

data Arc

Represents an arc with center, radius, ccw from start to end. note that angles must be in (-pi, pi)

pattern Arc :: Point -> !Float -> !Float -> !Float -> Arc

svgPathDefinition :: Arc -> String

Calculate the SVG path defition for an Arc

asTuple :: Arc -> (Float, Float, Float, Float, Float)

Convert a circle to a 2-tuple representation.

midpoint :: Arc -> Float -> Point

arcLength :: Arc -> Float

invert :: Arc -> Point -> Point -> Point

Reflect a point p about the arc at point b.